Correct strokes per interval, not per set
A set is a run of equal distances, not a promise of one stroke: four 50s
freestyle then three breaststroke is one set and two strokes, and labelling per
set made that correction unsayable and the wrong answer unfixable. Every
interval now carries its own dropdown, with the set control demoted to a bulk
setter for when it really was all one thing.

Two display bugs the interval view exposed. 'mixed' compared lengths, so one
undetermined length in a 1250 flagged the whole set as two strokes; it now
compares intervals. And a medley interval showed 'backstroke' - the same
four-way-tie-broken-alphabetically trap as the set label, one level down.

Also silences a pandas downcasting deprecation: a pool_length_m column that is
entirely NULL arrives as object dtype and fillna quietly downcast it.

+++ polarswim/analyze.py
@@ -117,7 +117,11 @@ def load_lengths(engine: Engine, workout_id: int | None = None) -> pd.DataFrame:
         return df
 
     derived = df["distance_m"] / df["n_lengths"].replace(0, np.nan)
-    df["pool_m"] = df["pool_length_m"].fillna(derived)
+    # Both cast to float first: where every pool_length_m is NULL the column
+    # arrives as object dtype, and filling it then silently downcasts — which
+    # pandas deprecates and will stop doing.
+    df["pool_m"] = (df["pool_length_m"].astype(float)
+                    .fillna(derived.astype(float)))
     df = df[df["pool_m"].notna() & (df["pool_m"] > 0)].copy()
     df["pace_s"] = df["duration_s"] * (REFERENCE_LENGTH_M / df["pool_m"])
     return df
